% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockR-package.R
\docType{data}
\name{df}
\alias{df}
\title{Digital Elevation Model based on the NASA STRM Mission, for the Hamilton, ON Area.}
\format{
A data frame with 3960031 rows and 3 variables:
\describe{
  \item{x}{Longitude}
  \item{y}{Latitude}
  \item{DEM_elevation_Hamilton}{Estimated height from sea level in meters}
}
}
\source{
Google Earth Engine Data Catalog - NASADEM: NASA NASADEM Digital Elevation 30m (https://developers.google.com/earth-engine/datasets/catalog/NASA_NASADEM_HGT_001) Accessed: November 7, 2021
}
\usage{
data(df)
}
\description{
It contains 3960031 georeference centroids, each with the height estimated by
the SRTM mission. More details at:
https://developers.google.com/earth-engine/datasets/catalog/NASA_NASADEM_HGT_001
}
\examples{
 data(df)

}
\keyword{datasets}
