% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_changes.R
\name{fetch_changes}
\alias{fetch_changes}
\title{Fetch Historical Exchange Rates from Yahoo Finance}
\usage{
fetch_changes(from, to)
}
\arguments{
\item{from}{A character string representing the base currency (e.g., "USD").}

\item{to}{A character string representing the target currency (e.g., "EUR").}
}
\value{
A dataframe with daily exchange rates. The returned dataframe contains the following columns:
\describe{
\item{timestamp}{int, the timestamp of the rate.}
\item{rate.high}{Numeric, the highest exchange rate of the day.}
\item{rate.low}{Numeric, the lowest exchange rate of the day.}
\item{rate.open}{Numeric, the opening exchange rate of the day.}
\item{rate.volume}{Numeric, the trading volume for the day.}
\item{rate.close}{Numeric, the closing exchange rate of the day.}
\item{date}{POSIXct, the corresponding date (YYYY-MM-DD).}
}
}
\description{
This function retrieves historical exchange rates for a given currency pair
from Yahoo Finance and returns a dataframe with daily exchange rate data
(values are supposed to be each precise minute of the same day).
}
\details{
The function also returns an additional \code{insights} \code{data.frame} as attributes.
Some of the key attributes include latest informations from Yahoo Finance API :
\describe{
\item{currency}{Character, the base currency.}
\item{symbol}{Character, the Yahoo Finance symbol (e.g., "EURUSD=X").}
\item{exchange_name}{Character, the exchange place name.}
\item{regular_market_price}{Numeric, the latest market price.}
\item{regularmarketdayhigh}{Numeric, the market highest price of this day.}
\item{regularmarketdaylow}{Numeric, the market lowest price of this day.}
\item{fifty_two_week_high}{Numeric, the highest price in the last 52 weeks.}
\item{fifty_two_week_low}{Numeric, the lowest price in the last 52 weeks.}
\item{previous_close}{Numeric, the last closing price.}
\item{timezone}{Character, the market's timezone.}
\item{range}{Character, the data range (e.g., "1d", "5d", "1mo").}
}
}
\examples{
# Fetch exchange rate data for USD to EUR
df <- fetch_changes("USD", "EUR")

# View the first few rows of the dataframe
head(df)

# Access metadata attributes
attr(df, "insights")
}
