% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_historic.R
\name{fetch_historic}
\alias{fetch_historic}
\title{Fetch historical financial data indices}
\usage{
fetch_historic(symbols = c("SAAB-B.ST"), wait.time = 0, .verbose = T, ...)
}
\arguments{
\item{symbols}{\code{character} A character string representing the financial indices to search for, e.g., ticker symbol(s).}

\item{wait.time}{\code{double}, default = \code{0} A character string representing an additional waiting time between 2 calls to the Yahoo API.}

\item{.verbose}{\code{logical}, default = \code{TRUE}. If \code{TRUE}, send messages to the console.}

\item{...}{
  Arguments passed on to \code{\link[=get_yahoo_data]{get_yahoo_data}}
  \describe{
    \item{\code{symbol}}{A character string representing the symbol for the financial instrument (e.g., "AAPL" for Apple).}
    \item{\code{start_date}}{A character string representing the start date in the format "YYYY-MM-DD". If \code{NULL}, data starts from 1970-01-01.}
    \item{\code{end_date}}{A character string representing the end date in the format "YYYY-MM-DD". If \code{NULL}, data is retrieved up to the current date.}
    \item{\code{interval}}{A character string representing the interval for the returned datas : default will return daily values.
Other intervals will filter out some of these daily values, depending on the desired interval.
Valid interval are "1d", "5d", "1mo", "3mo", "6mo", "1y", "2y", "5y", "10y", "ytd", and "max".}
  }}
}
\value{
A data frame containing the historical financial data with the following columns:
\item{open}{\code{numeric} The opening price for the period (default is each day).}
\item{close}{\code{numeric} The closing price for the period (default is each day).}
\item{adjclose}{\code{numeric} The adjusted closing price on the period, which accounts for corporate actions like dividends and stock splits.}
\item{low}{\code{numeric} The lowest price for the period (default is each day).}
\item{high}{\code{numeric} The highest price for the period (default is each day).}
\item{volume}{\code{integer} The traded volume.}
\item{timestamp}{\code{integer} Unix timestamps corresponding to each data point.}
\item{date}{\code{POSIXct} The day of the financial data point.}
\item{currency}{\code{character} The currency in which the data is reported, depending on the marketplace.}
\item{symbol}{\code{character} The stock or financial instrument symbol (e.g., "AAPL").}
\item{shortname}{\code{character} The abbreviated name of the company or financial instrument.}
\item{longname}{\code{character} The full name of the company or financial instrument.}
\item{exchangename}{\code{character} The name of the exchange marketplace where the financial instrument is listed.}
\item{fullexchangename}{\code{character} The full name of the exchange marketplace.}
\item{timezone}{\code{character} The timezone in which the data is reported.}
}
\description{
Given indices names (i.e. ticker symbol), the function retrieves historical stock market data
from finance.yahoo.com. Answer a \code{data.frame} with standardized colnames.
}
\details{
Return a \code{data.frame} (see returned columns in 'Value' section) with additional attributes :
\itemize{
\item \code{fetch.symbols}: the symbols originally asked by the user
\item \code{fetch.date}: the date when data are retrieved
\item \code{fetch.currencies}: the currencies within the \code{data.frame}
\item \code{n.currencies}: the number of currencies within the \code{data.frame}
}
}
\examples{
datas <- fetch_historic(symbols = c("VOLCAR-B.ST", "SAAB-B.ST") )

}
\seealso{
\code{\link{get_yahoo_data}}
}
