% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_stock_indices.R
\name{fetch_stock_indices}
\alias{fetch_stock_indices}
\title{Fetch stock indices based on a company name}
\usage{
fetch_stock_indices(names, marketplaces = NULL)
}
\arguments{
\item{names}{A character string representing the company name to search for.}

\item{marketplaces}{(optionnal) A character string representing the marketplace(s) to consider. Default keep all the marketplace}
}
\value{
A data frame with columns:
- \code{Symbol}: The stock ticker symbol from yahoo
- \code{Name}: The full company name.
- \verb{Last Price}: The latest available price.
- \code{Sector / Category}: The sector or industry category (if available).
- \code{Type}: The type of asset (always "stocks").
- \code{Exchange}: The stock exchange place for this stock.
\itemize{
\item \code{searched}: The original names searched
}
}
\description{
Given companies names, the function retrieves overall stock market data
(by searching on https://finance.yahoo.com ).
It returns a data frame with the stock symbol on various marketplaces
, companies names, last price on the marketplace,
sector/category (if available), type (always "stocks")
, exchange marketplace name and initially searched companies names.
}
\examples{
fetch_stock_indices(names = c("VOLVO", "SAAB"),  marketplaces = "STO"  )
}
