% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotiff_to_rda.R
\name{geotiff_to_rda}
\alias{geotiff_to_rda}
\title{Transform GeoTIFF file into tabular RDA file}
\usage{
geotiff_to_rda(input_url, output_url)
}
\arguments{
\item{input_url}{Full url (with filename and extension) for the GeoTIFF file}

\item{output_url}{Full url (with filename and extension *.rda) for the RDA file to be created}
}
\value{
The conversion of a GeoTIFF image to an RDA file
}
\description{
Transform GeoTIFF file into tabular RDA file
}
