% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_changes.R
\name{get_changes}
\alias{get_changes}
\title{Get Exchange Rates from Yahoo Finance, Given Devises}
\usage{
get_changes(from = NULL, to = "USD")
}
\arguments{
\item{from, }{default = \code{NULL} A character string representing the base currency (e.g., "USD").}

\item{to, }{default = \code{"USD"} A character string representing the target currency (e.g., "EUR").}
}
\value{
A dataframe with daily exchange rates. The returned dataframe contains the following columns:
\describe{
\item{timestamp}{int, the timestamp of the rate.}
\item{rate.high}{Numeric, the highest exchange rate of the day.}
\item{rate.low}{Numeric, the lowest exchange rate of the day.}
\item{rate.open}{Numeric, the opening exchange rate of the day.}
\item{rate.volume}{Numeric, the trading volume for the day.}
\item{rate.close}{Numeric, the closing exchange rate of the day.}
\item{date}{POSIXct, the corresponding date (YYYY-MM-DD).}
}
}
\description{
This function retrieves exchange rates for a given currency pair. Default convert to USD ($)
Source is Yahoo Finance API. It returns a \code{data.frame} with daily exchange rate data.
Returned \code{data.frame} is historical values on a day and latest insights within attributes
Returned values are supposed to be each minute within a day, more or less (≃ 1200 obs. ± 200 obs.).
}
\details{
The function also returns an additional \code{insights} \code{data.frame} as attributes.
Some of the key attributes include latest informations from Yahoo Finance API :
\describe{
\item{currency}{\code{character}, the base currency.}
\item{symbol}{\code{character}, the Yahoo Finance symbol (e.g., "EURUSD=X").}
\item{exchange_name}{\code{character}, the exchange place name, i.e. CCY for currencies.}
\item{regularmarketprice}{\code{numeric}, the latest market price.}
\item{regularmarketdayhigh}{\code{numeric}, the market highest price of this day.}
\item{regularmarketdaylow}{\code{numeric}, the market lowest price of this day.}
\item{fifty_two_week_high}{\code{numeric}, the highest price in the last 52 weeks.}
\item{fifty_two_week_low}{\code{numeric}, the lowest price in the last 52 weeks.}
\item{previous_close}{\code{numeric}, the last closing price.}
\item{timezone}{\code{character}, the market's timezone.}
\item{date}{\code{POSIXct} Equivalent to the timestamp of \code{regularmarketprice}, date of the observation.}
\item{from}{\code{character}, the currency converted into another, e.g., if the \code{from} value is 1$ ('USD'), you want to receive a certain amount of the other currency to reach 1$.}
\item{to}{\code{character}, the currency that you want to convert into : \strong{all the \code{numeric} values (not \code{integer}) in this line of the \code{data.frame} are expressed with this currency}.}
}
}
\examples{
# Fetch exchange rate data for USD to EUR
df <- get_changes(from = c("USD", "EUR"), to = "RON")
# Convert USD & EUR to Romanian Leu (RON)
head(df)
# or pass like a named list of character
df2 <- get_changes(from = c("EUR" = "RON", "USD" = "EUR"))
# Or indicate paired values as 2 list :
same_as_df2 <- get_changes(from = c("EUR", "USD"), c("RON" , "EUR"))
# Access metadata attributes
attr(df, "insights")
}
