% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_currencies.R
\name{get_currencies}
\alias{get_currencies}
\title{Fetch Financial currencies Names}
\usage{
get_currencies(keep = NULL)
}
\arguments{
\item{keep}{A character vector of symbols to filter the results. If NULL (default),
no filtering is applied, and all available indices are returned.}
}
\value{
A data frame containing unique financial indices (currencies). The table has
columns like \code{symbol}, \code{name}, and other relevant information, with all column names in lowercase.
If \code{keep} is specified, only the matching indices are returned.
}
\description{
This function fetches a table of financial indices (currencies) from Yahoo Finance.
Optionally, it can filter the results to include only the specified indices.
}
\details{
The function sends a request to Yahoo Finance's API to fetch a list of available currencies.
It then processes the results and returns them as a data frame. If an internet connection is not available,
a warning message is shown.
}
\examples{
# Fetch all available indices
all_indices <- get_currencies()

# Fetch only specific indices
selected_indices <- get_currencies(keep = c("USD", "EUR"))

}
