% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indices.R
\name{get_indices}
\alias{get_indices}
\title{Retrieve Stock Market Indices from Yahoo Finance}
\usage{
get_indices()
}
\value{
A data frame with the following columns:
\item{symbol}{Character. The ticker symbol of the index (e.g., \verb{^GSPC} for S&P 500).}
\item{name}{Character. The full name of the index (e.g., "S&P 500").}
\item{price}{\code{numeric}  The current value of the index.}
\item{change}{\code{numeric} The absolute change in index value.}
\item{change_percent}{\code{numeric} The percentage change in index value.}
}
\description{
This function fetches real-time stock market indices, including price, change, and percentage change, from Yahoo Finance.
}
\examples{
\dontrun{
indices <- get_indices()
head(indices)
}

}
