% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yahoo_data.R
\name{get_yahoo_data}
\alias{get_yahoo_data}
\title{Get Yahoo Finance Data for a Given Symbol}
\usage{
get_yahoo_data(
  symbol = "AAPL",
  start_date = NULL,
  end_date = NULL,
  interval = "1d"
)
}
\arguments{
\item{symbol}{A character string representing the symbol for the financial instrument (e.g., "AAPL" for Apple).}

\item{start_date}{A character string representing the start date in the format "YYYY-MM-DD". If \code{NULL}, data starts from 1970-01-01.}

\item{end_date}{A character string representing the end date in the format "YYYY-MM-DD". If \code{NULL}, data is retrieved up to the current date.}

\item{interval}{A character string representing the interval for the returned datas : default will return daily values.
Other intervals will filter out some of these daily values, depending on the desired interval.
Valid interval are "1d", "5d", "1mo", "3mo", "6mo", "1y", "2y", "5y", "10y", "ytd", and "max".}
}
\value{
A data frame containing the historical financial data with the following columns:
\item{open}{\code{numeric} The opening price for the period (default is each day).}
\item{close}{\code{numeric} The closing price for the period (default is each day).}
\item{adjclose}{\code{numeric} The adjusted closing price on the period, which accounts for corporate actions like dividends and stock splits.}
\item{low}{\code{numeric} The lowest price for the period (default is each day).}
\item{high}{\code{numeric} The highest price for the period (default is each day).}
\item{volume}{\code{integer} The traded volume.}
\item{timestamp}{\code{integer} Unix timestamps corresponding to each data point.}
\item{date}{\code{POSIXct} The day of the financial data point.}
\item{currency}{\code{character} The currency in which the data is reported, depending on the marketplace.}
\item{symbol}{\code{character} The stock or financial instrument symbol (e.g., "AAPL").}
\item{shortname}{\code{character} The abbreviated name of the company or financial instrument.}
\item{longname}{\code{character} The full name of the company or financial instrument.}
\item{exchangename}{\code{character} The name of the exchange marketplace where the financial instrument is listed.}
\item{fullexchangename}{\code{character} The full name of the exchange marketplace.}
\item{timezone}{\code{character} The timezone in which the data is reported.}
}
\description{
Given the symbol of a financial value, this function retrieves historical financial data from Yahoo Finance
and answer a \code{data.frame} of historical values.
The data includes open stock price, high, low, close, volume, and adjusted close prices, along with timestamps (i.e. the day associated with the measures).
}
\details{
The default \code{data.frame} have a line for each day.
If the user provide another interval range than '1d' (one day), lines will be filtered out, in order to match the desired interval range.
Valid interval ranges are: "1d", "5d", "1mo", "3mo", "6mo", "1y", "2y", "5y", "10y", "ytd", and "max".

The function allows the user to specify a date range using start and end dates. If no date range is specified,
it retrieves all available data from the beginning of time (default for start) to the current date (default for end).
}
\examples{

  data <- get_yahoo_data(symbol = "SAAB-B.ST", start_date = "2020-01-01", interval = "5d")
  head(data)

}
