% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockAnalyzer.R
\docType{class}
\name{StockAnalyzer-class}
\alias{StockAnalyzer-class}
\title{StockAnalyzer
The main controller for stockr package.
It automatically download historical OHLC data from a specific CSV file,
and calculate the performance of the portfolio.}
\description{
StockAnalyzer
The main controller for stockr package.
It automatically download historical OHLC data from a specific CSV file,
and calculate the performance of the portfolio.
}

