% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockR-package.R
\docType{package}
\name{stockR-package}
\alias{stockR}
\alias{stockR-package}
\title{stockR: Provides resources for carbon stock modelers}
\description{
For now, it provides a Digital Elevation Model for the region of Hamilton, Ontario, and a function to convert GeoTIFF files to RDA. Further capabilities will be added soon.

This package is an exercise in the class of prof. Antonio Paez.
}
\references{
\url{https://github.com/ricardobarroslourenco/stockR}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ricardobarroslourenco/stockR}
}

}
\author{
\strong{Maintainer}: Ricardo Barros Lourenco \email{barroslr@mcmaster.ca} (\href{https://orcid.org/0000-0002-4158-3244}{ORCID})


Ricardo Barros Lourenco, School of Earth, Environment and Society, McMaster University \email{barroslr@mcmaster.ca}
}
\keyword{internal}
