% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_historic.R
\name{fetch_historic}
\alias{fetch_historic}
\title{Fetch historical financial data indices}
\usage{
fetch_historic(symbols = c("SAAB-B.ST"), wait.time = 0, .verbose = T)
}
\arguments{
\item{symbols}{\code{character} A character string representing the indices to search for.}

\item{wait.time}{\code{double} A character string representing the indices to search for.}

\item{.verbose}{\code{logical}, default = \code{TRUE}. If \code{TRUE}, send messages to the console}
}
\value{
A data frame with columns:
- \code{low}: Lowest value of the indice during this day.
- \code{high}: Highest value of the indice during this day.
- \code{close}: Value at the closing of the maketplace.
- \code{open}: The value at the opening of the marketplace
- \code{volume}: The volume of operation.
- \code{adjclose}: The adjusted closing value.
\itemize{
\item \code{timestamp}: The date for these values
\item \code{date}: The date for these values
\item \code{symbol}: The original indices names searched
\item \code{shortname}: The name of the value
\item \code{exchangename}: The financial place abbreviation
\item \code{fullexchangename}: The full name of the financial place
\item \code{timezone}: The timezone, e.g., 'CET' for Central European Time
}
}
\description{
Given indices names, the function retrieves historical stock market data
(from finance.yahoo.com by default). Answer a \code{data.frame} with standardized colnames.
}
\examples{
datas <- fetch_historic(symbols = c("VOLCAR-B.ST", "SAAB-B.ST") )
}
\seealso{
\code{\link{get_yahoo_data}}
}
