% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_historic.R
\name{fetch_historic}
\alias{fetch_historic}
\title{Fetch stock indices with \code{quantmod}}
\usage{
fetch_historic(
  symbols = c("SAAB-B.ST"),
  wait.time = 0,
  .verbose = T,
  src = "yahoo"
)
}
\arguments{
\item{symbols}{\code{character} A character string representing the indices to search for.}

\item{wait.time}{\code{double} A character string representing the indices to search for.}

\item{.verbose}{\code{logical}, default = \code{TRUE}. If \code{TRUE}, send messages to the console}

\item{src}{\code{character}, default = \code{'yahoo'} The src parameters of  quantmod::getSymbols.}
}
\value{
A data frame with columns:
- \code{open}: The value at the opening of the marketplace
- \code{high}: Highest value of the indice during this day.
- \code{low}: Lowest value of the indice during this day.
- \code{close}: Value at the closing of the maketplace.
- \code{volume}: The volume of operation.
- \code{adjusted}: The adjusted closing value.
\itemize{
\item \code{symbol}: The original indices names searched
\item \code{date}: The date for these values
}
}
\description{
Given indices names, the function retrieves historical stock market data
(from finance.yahoo.com by default).
It returns the data frame returned by quantmod::getSymbols
with a proper date column, standardized colnames.
}
\examples{
datas <- fetch_historic(symbols = c("VOLCAR-B.ST", "SAAB-B.ST") )
}
\seealso{
\url{https://www.quantmod.com/documentation/getSymbols.html}
}
