% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_historic.R
\name{fetch_historic}
\alias{fetch_historic}
\title{Fetch historical financial data indices}
\usage{
fetch_historic(symbols = c("SAAB-B.ST"), wait.time = 0, .verbose = T, ...)
}
\arguments{
\item{symbols}{\code{character} A character string representing the indices to search for.}

\item{wait.time}{\code{double} A character string representing the indices to search for.}

\item{.verbose}{\code{logical}, default = \code{TRUE}. If \code{TRUE}, send messages to the console.}

\item{...}{Parameters passed to \code{get_yahoo_data} such as the \code{interval} between two lines, a start_date or end_date}
}
\value{
A data frame with columns:
- \code{low}: Lowest value of the indice during this day.
- \code{high}: Highest value of the indice during this day.
- \code{close}: Value at the closing of the maketplace.
- \code{open}: The value at the opening of the marketplace
- \code{volume}: The volume of operation.
- \code{adjclose}: The adjusted closing value.
\itemize{
\item \code{timestamp}: The date for these values
\item \code{date}: The date for these values
\item \code{symbol}: The original indices names searched
\item \code{shortname}: The name of the value
\item \code{exchangename}: The financial place abbreviation
\item \code{fullexchangename}: The full name of the financial place
\item \code{timezone}: The timezone, e.g., 'CET' for Central European Time
}

This \code{data.frame} have additional attributes :
\itemize{
\item \code{fetch.symbols}: the symbols originally asked by the user
\item \code{fetch.date}: the date when data are retrieved
\item \code{fetch.currencies}: the currencies within the \code{data.frame}
\item \code{n.currencies}: the number of currencies within the \code{data.frame}
}
}
\description{
Given indices names (i.e. ticker symbol), the function retrieves historical stock market data
from finance.yahoo.com. Answer a \code{data.frame} with standardized colnames.
}
\examples{
datas <- fetch_historic(symbols = c("VOLCAR-B.ST", "SAAB-B.ST") )
}
\seealso{
\code{\link{get_yahoo_data}}
}
