% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_historic.R
\name{fetch_historic}
\alias{fetch_historic}
\title{Fetch historical financial data indices}
\usage{
fetch_historic(symbols = c("SAAB-B.ST"), wait.time = 0, .verbose = T, ...)
}
\arguments{
\item{symbols}{\code{character} A character string representing the indices to search for.}

\item{wait.time}{\code{double} A character string representing the indices to search for.}

\item{.verbose}{\code{logical}, default = \code{TRUE}. If \code{TRUE}, send messages to the console.}

\item{...}{Parameters passed to \code{get_yahoo_data} such as the \code{interval} between two lines, a start_date or end_date}
}
\value{
A data frame containing the historical financial data with the following columns:
\item{open}{\code{numeric} The opening price for the period (default is each day).}
\item{close}{\code{numeric} The closing price for the period (default is each day).}
\item{adjclose}{\code{numeric} The adjusted closing price on the period, which accounts for corporate actions like dividends and stock splits.}
\item{low}{\code{numeric} The lowest price for the period (default is each day).}
\item{high}{\code{numeric} The highest price for the period (default is each day).}
\item{volume}{\code{integer} The traded volume.}
\item{timestamp}{\code{integer} Unix timestamps corresponding to each data point.}
\item{date}{\code{POSIXct} The day of the financial data point.}
\item{currency}{\code{character} The currency in which the data is reported, depending on the marketplace.}
\item{symbol}{\code{character} The stock or financial instrument symbol (e.g., "AAPL").}
\item{shortname}{\code{character} The abbreviated name of the company or financial instrument.}
\item{longname}{\code{character} The full name of the company or financial instrument.}
\item{exchangename}{\code{character} The name of the exchange marketplace where the financial instrument is listed.}
\item{fullexchangename}{\code{character} The full name of the exchange marketplace.}
\item{timezone}{\code{character} The timezone in which the data is reported.}
}
\description{
Given indices names (i.e. ticker symbol), the function retrieves historical stock market data
from finance.yahoo.com. Answer a \code{data.frame} with standardized colnames.
}
\details{
This \code{data.frame} have additional attributes :
\itemize{
\item \code{fetch.symbols}: the symbols originally asked by the user
\item \code{fetch.date}: the date when data are retrieved
\item \code{fetch.currencies}: the currencies within the \code{data.frame}
\item \code{n.currencies}: the number of currencies within the \code{data.frame}
}
}
\examples{
datas <- fetch_historic(symbols = c("VOLCAR-B.ST", "SAAB-B.ST") )

}
\seealso{
\code{\link{get_yahoo_data}}
}
