% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_indices.R
\name{fetch_indices}
\alias{fetch_indices}
\title{Fetch ticker stock indices based on a company name}
\usage{
fetch_indices(names, marketplaces = NULL)
}
\arguments{
\item{names}{A character string representing the company name to search for.}

\item{marketplaces}{(optionnal) A character string representing the marketplace(s) to consider. Default keep all the marketplace}
}
\value{
A data frame with columns:
- \code{symbol}: The stock ticker symbol from yahoo
- \code{name}: The full company name.
- \code{last_price}: The latest available price.
- \code{sector_category}: The sector or industry category (if available).
- \code{type}: The type of asset (always "stocks").
- \code{exchange}: The stock exchange place for this stock.
\itemize{
\item \code{searched}: The original names searched
}
}
\description{
Given companies names, the function retrieves overall stock market data
(by searching on https://finance.yahoo.com ).
It returns a data frame with the ticker symbol on various marketplaces
, companies names, last price on the marketplace,
sector/category (if available), type (always "stocks")
, exchange marketplace name and initially searched companies names.
}
\examples{
fetch_indices(names = c("VOLVO", "SAAB"),  marketplaces = "STO")
}
\seealso{
\code{\link{get_yahoo_data}},  \code{\link{fetch_historic}}
}
